Remove the Context concept, add/fix some executables.
The anticipated need for Context() was based on the idea that references
would be resolved prior to the XML being built. This demanded that the intermediate
structure be queried which was beyond its desired use case. XPath is a much better
solution for this particular problem as the ultimate structure could be something
that cannot be anticipated upstream.