Fix test_serialize.
Validate exit operations properly and manage the Sector's context.

Serialize was in good condition from the manual testing, so this is compensation for some changes.
The only difference was that we aren't working with layer subclasses in the test, so change
terminal checks to not depend on the availability of the terminal attribute.