Improve race condition detection of exit events.
In some cases, tracked processes weren't getting exit events. It was likely
that macos was not emitting exit events when the kqueue parent added the process
after the child exited (hanging parent waiting for exit of zombie process). To counter,
the process is checked for exit after the process it tracked and removed if it did.

Also, the ability to attach data was removed. Rather, the connect mapping should just be
used as it currently it.

! NOTE:
	This logic needs to be moved to system context. Subprocess may be the user only,
	but there's no guarantee.