Fully delegate serialization.
Allows for custom tree builders to be used instead of XML serialization.
(soon to be used for project.txt files)

Update to use the serializer located in fault.web instead of the deprecating
fault.xml.

Use class local aliases for producing node events.

Begin renaming "literals" to "syntax" in order to avoid confusion with
paragraph "literal".