Add wait limit parameter for I/O transaction timeout.
Originally, the intention for withholding access to waittime was
to encourage use in a distinct thread to properly leverage the parallel I/O.
The timeout being managed would be handled by the receiving queue.

However, this adds undesireable complexity to certain applications, so implement
the wait method which will likely be the standard way to access the I/O transaction
moving forward: `with array.wait(limit) as xact` as opposed to `with array as xact`.