Eliminate template hosts.
Given the capacity to load multiple host files, splitting hosts across
the files with the relevant context should be sufficient for most cases.

The :Inherit directive was more powerful, but forced awkward configuration
processing. Either a postprocess to apply inheritance or making sure
the host index was available during parsing to resolve the dependency. The
implementation had chosen the latter requiriing that template hosts were
already defined at the time their reference was parsed.

The context host is similar, but the unconditional application allows better
factoring of the parsing functions and simplifies the http-network conceptually.

In more advanced cases, an actual templating engine should be leveraged
to generate the configuration file.