Implement Test.itertimer for measuring performance.
Extend Test to allow the capture of arbitrary metrics to accommodate for
the timings, `iterations` and `timer`. Additionally, track contentions
automatically so that symmetry can be provided in cases where performance
information is expected by a data channel--`contentions` and `duration`
instead of `iterations` and `timer`.

There is some likelihood of metrics being a mere vector of count-time
pairs. For the moment, leave it as an open dictionary.