Add trapped reports and configure METRICS_IDENTITY for project metrics.
When factors are built for development purposes, a metrics trap argument
is identified by the compilation context. Use this parameter to identify
a location to store test reports--normally build cache relative.
Currently, this is more detailed than what is provided via status frames.
Future versions will have full alignment as some applications will prefer
to capture the data via frames(sentries).

The configuration of METRICS_IDENTITY and PROCESS_IDENTITY is provided
for trapped builds that use parallel channels to store the trace data.
When a factor built for data capture is executed, only the factor is
known by the instrumented code. So, it is necessary to provide the
identity of the application to allow for per-test data captures(coverage).

With the detailed test reports, improve the resolution of the test's
resource usage.