Implement .project declarations to supersede .protocol and project.txt.
Rework the factor protocol declaration to allow the identity of a project directory
with a mere file test and to include more information. With `.protocol` being
generic, reading the file was necessary to recognize the meaning of the directory.

While `.protocol` allows for easier out-of-context scanning, the context of a scan
is normally known and, in some cases, it may be necessary to have parallel semantics.
For most operations, products, projects, or factors are being scanned for, but rarely
are the scans looking for any or all of the three. For parallel semantics, where a directory
is serving as product or project and factor, `.protocol` is insufficient without changes.

The additional information fields, over .protocol, are intended to replace `project.txt`
entirely. Abstract and Icons being ignored as they are not immediately necessary at
that this level. &.types.Information will eventually drop icons and abstract entirely.