Implement Phrase.seek and remove the Units string container.
Seek, in some form, will replace most of the scanning methods on Phrase once
a more reasonable means of unit selection is identified(Multiple implementations of the
same interface on Words[cell, unit, codepoint]?)

The use of the 'unit' trait is likely temporary as an explicit Words type and subclass
should provide a less irrelevant location to designate a UPC. The removed Units
was not entirely appropriate as it allowed for multiple distinct character units leading
to a possible inconsistency in the cell usage rate of the string.

An early step toward cleaning up the redundant methods in Phrase and Character Unit handling.