Remove old cell tearing interfaces and Context.print.
&.types.Phrase.seek, &.types.Phrase.tell, and &.matrix.Context.view
will, eventually, provide the desired functionality. Implementing
tearing at the Phrase or Word level is not appropriate as the display
Context may support proper partial display. (Likely non-legacy tty displays)

&.matrix.Context.print was removed in favor of manual seeks and renders.
With differential rendering, full page prints are not that common and it
may not be clear whether or not Context.view should be applied or not,
so force the user to resolve the ambiguity by implementing print themselves.