Transition to product relative image directories.
For polynomial, storing images relative to the factor sources meant that
the isolation protocol needed to be aware of the image directory and even define
where images were stored. This is problematic as it requires an implementation of
the protocol to be loaded in order to find images. With an integrated, system defined,
protocol, no such resolution is required making factor execution a much simpler
procedure as an abstraction is eliminated.

The value loss from this change is that moving projects or context directories no
longer implies the carrying of the images. However, whether or not that value
is present depends on the system's integration process. If paths are hard coded
into images, the value was not there to begin with. For some library factors,
this is already the case with the current implementation.

Currently, form is essentially ignored, but will return to house delineation
and metrics data or any builds that should sit parallel with execution images.