Remove the terminal emulation abstraction project.
Primarily developed for simple formatting assistance and to allow
terminal.syntax' original implementation to work within terminal emulators,
terminal's primary function was to provide more accessible identities to
common DEC/ANSI escapes. While functional, the difficulty of testing it
across terminal emulator implementations left it in a continuous state of
uncertainty. In part, this and inconsistencies in emulators led to the
development of terminal/io and, subsequently, the removal of python/terminal as
terminal/syntax was the primary dependent.