Revise TTYN.1 to use data URLs for packing frame data extensions.
While the selected OSC "code" had the misfortune of conflicting with modern terminal emulators,
it is an opportunity for an improvement.

OSC-8 allows URLs to be associated with displayed cells; solving a data accessibility problem
with frame extensions. Emulators providing access to these URLs will be able to open or copy
the extension contents without the use or knowledge of another tool by the user.

- Eliminate compression to simplify frame I/O and allow the data URL to function without
. surprising mime type configurations. (Better performance is also likely with channel compression).
- Adjust the signal and size color to something other than green. (dim blue)
. Preferring to reserve greens, yellows, and reds to exclusively represent specific information.
- Prefer the `http://fault.io/protocol/` prefix.